+++ pkg/scad/transformation/color.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package transforms
+package transformation
 
 import (
 	"github.com/micahkemp/scad/pkg/scad/values"
