+++ pkg/scad/transformation/hull.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package transforms
+package transformation
 
 // Hull is a hull transform.
 type Hull struct {
