+++ pkg/scad/transformation/minkowski.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package transforms
+package transformation
 
 // Minkowski is a minkowsi transform.
 type Minkowski struct {
